Feat: Testimony and user validation with YUP

+++ src/services/TestimonialService.ts
@@ -1,8 +1,4 @@
-import {
-  ITestimonial,
-  ITestimonialCreate,
-  ITestimonialUpdate,
-} from '../models/Testimonial'
+import { Testimonial } from '@prisma/client'
 import { TestimonialRepository } from '../repository/TestimonialRepository'
 
 export class TestimonialService {
@@ -12,32 +8,42 @@ export class TestimonialService {
     this.testimonialRepository = new TestimonialRepository()
   }
 
-  async create(
-    testimonialData: ITestimonialCreate,
-  ): Promise<ITestimonialCreate> {
+  async create(testimonialData: Testimonial): Promise<Testimonial> {
     return await this.testimonialRepository.create(testimonialData)
   }
 
-  async getAll(): Promise<ITestimonial[]> {
+  async getAll(): Promise<Testimonial[]> {
     return await this.testimonialRepository.getAll()
   }
 
-  async getById(id: string): Promise<ITestimonial | null> {
-    return await this.testimonialRepository.getById(id)
+  async getById(id: string): Promise<Testimonial | null> {
+    const testimonial = await this.testimonialRepository.getById(id)
+
+    if (!testimonial) {
+      throw new Error('Depoimento não encontrado')
+    }
+
+    return testimonial
   }
 
-  async getRandom(): Promise<ITestimonial[]> {
+  async getRandom(): Promise<Testimonial[]> {
     return await this.testimonialRepository.getRandom()
   }
 
   async update(
     id: string,
-    testimonialData: ITestimonialUpdate,
-  ): Promise<ITestimonialUpdate | null> {
+    testimonialData: Testimonial,
+  ): Promise<Testimonial | null> {
     return await this.testimonialRepository.update(id, testimonialData)
   }
 
-  async delete(id: string): Promise<ITestimonial | null> {
+  async delete(id: string): Promise<Testimonial | null> {
+    const testimonial = await this.testimonialRepository.getById(id)
+
+    if (!testimonial) {
+      throw new Error('Depoimento não encontrado')
+    }
+
     return await this.testimonialRepository.delete(id)
   }
 }
