+++ src/services/UserService.ts
@@ -1,4 +1,4 @@
-import { IUser, IUserCreate, IUserUpdate } from '../models/User'
+import { User } from '@prisma/client'
 import { UserRepository } from '../repository/UserRepository'
 
 export class UserService {
@@ -8,23 +8,35 @@ export class UserService {
     this.userRepository = new UserRepository()
   }
 
-  async create(userData: IUserCreate): Promise<IUserCreate> {
+  async create(userData: User): Promise<User> {
     return await this.userRepository.create(userData)
   }
 
-  async getAll(): Promise<IUser[]> {
+  async getAll(): Promise<User[]> {
     return await this.userRepository.getAll()
   }
 
-  async getById(id: string): Promise<IUser | null> {
-    return await this.userRepository.getById(id)
+  async getById(id: string): Promise<User | null> {
+    const user = await this.userRepository.getById(id)
+
+    if (!user) {
+      throw new Error('Usuário não encontrado')
+    }
+
+    return user
   }
 
-  async update(id: string, userData: IUserUpdate): Promise<IUserUpdate | null> {
+  async update(id: string, userData: User): Promise<User | null> {
     return await this.userRepository.update(id, userData)
   }
 
-  async delete(id: string): Promise<IUser | null> {
+  async delete(id: string): Promise<User | null> {
+    const user = await this.userRepository.getById(id)
+
+    if (!user) {
+      throw new Error('Usuário não encontrado')
+    }
+
     return await this.userRepository.delete(id)
   }
 }
